% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age_in_yrs.R
\name{calc_age_in_yrs}
\alias{calc_age_in_yrs}
\title{Calculate age in years.}
\usage{
calc_age_in_yrs(birth_dt, as_of_dt)
}
\arguments{
\item{birth_dt}{Input birthdate as a string with formatting YYYY/MM/DD (e.g., "1981/02/25")}

\item{as_of_dt}{Input date at which age is to be calculated as a string with formatting YYYY/MM/DD (e.g., "2019/05/21")}
}
\description{
This function calculates age in years as of a certain date, correctly accounting for leap years. You could also use it to calculate years between any two dates.
}
\examples{
\dontrun{
born <- "1981/02/25"
yesterday <- Sys.Date() - 1
calc_age_in_yrs(born, yesterday)
}
}
