% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refuns.R
\name{circle_fun}
\alias{circle_fun}
\title{Generate circle data.}
\usage{
circle_fun(center = c(0, 0), radius = 2, npoints = 100)
}
\arguments{
\item{center}{Numeric vector for the x and y position of the center of the circle. Default is c(0, 0).}

\item{radius}{Radius of the circle. Default is 2.}

\item{npoints}{The number of data points to generate. Default is 100.}
}
\description{
This function generates data that can be used to draw circles in ggplot.
}
\examples{
\dontrun{
circle_data <- circle_fun(c(1,-1), 2.3, npoints = 100)
library("ggplot2")
ggplot(dat,aes(x,y)) + geom_path()
}
}
