% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_all_pkgs.R
\name{detach_all_pkgs}
\alias{detach_all_pkgs}
\title{Detach all packages.}
\usage{
detach_all_pkgs()
}
\description{
This function detaches as many aspects of installed packages from the working environment as is possible without a restart of the R session. Note that in some packages, it is not possible to fully detach them without restarting R.
}
