% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_and_load_pkgs.R
\name{install_and_load_pkgs}
\alias{install_and_load_pkgs}
\title{(Install and) load packages.}
\usage{
install_and_load_pkgs(pkgs_needed)
}
\arguments{
\item{pkgs_needed}{Input a character vector of the packages you want to install/load. Can be a single package (e.g., "dplyr") or multiple (c("dplyr", "tidyr")).}
}
\description{
This function checks if each package that needs to be loaded has been installed, installs any that are not, and then loads all of the listed packages.
}
\examples{
\dontrun{
pkgs_needed <- c("tidyverse", "zoo", "openxlsx")
install_and_load_pkgs(pkgs_needed)
}
}
