% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isid.R
\name{isid}
\alias{isid}
\title{A Stata-esque function for asserting an id variable is unique.}
\usage{
isid(x)
}
\arguments{
\item{x}{Input the ID to be checked for duplicates. Can be a single variable, or multiple.}
}
\description{
Based on the Stata function of the same name, this function asserts whether rows of a dataframe are unique, and if they are not, it will throw an error and, when sourced, will stop the sourced script code from running any further.
}
