% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refuns.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert radians to degrees.}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{Input the radian that is to be converted into degrees.}
}
\description{
This function converts radians to degrees. It is commonly needed in code that we use to graph circles.
}
