% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refuns-package.R
\docType{package}
\name{refuns-package}
\alias{refuns}
\alias{refuns-package}
\title{refuns: A collection of functions used by the GCPS Office of Research & Evaluation}
\description{
The miscellaneous functions contained in this package are intended to make life a little easier when using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gcpsresearch/refuns}
  \item Report bugs at \url{https://github.com/gcpsresearch/refuns/issues}
}

}
\author{
\strong{Maintainer}: Matthew Lovelace \email{matthew.lovelace@gcpsk12.org}

Authors:
\itemize{
  \item James Appleton \email{james.appleton@gcpsk12.org}
}

}
\keyword{internal}
