% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refuns.R
\name{rnd_like_xl}
\alias{rnd_like_xl}
\title{Round like Excel.}
\usage{
rnd_like_xl(val, n_dec)
}
\arguments{
\item{val}{Input the value to be rounded.}

\item{n_dec}{Input the number of decimal places to round to (e.g., 0 = round to integer, 1 = round to the tenths place, 2 = round to the hundreths place, etc.).}
}
\description{
This function rounds numbers in the same way that Excel does:

Rounds up the number line when values are positive and trailing value is 5-9.

Rounds down the number line when values are negative and trailing value is 5-9.

If a trailing value is 5 exactly, always rounds in the same direction (i.e., up the number
line when value is positive, down the number line when value is negative). This is unlike
R which uses a more statistically valid but less widely accepted rounding rule of randomly
rounding half of the exact 5s up and half of the exact 5s down.

Solves floating point rounding errors that seem to occur at a rate of about 1/200 or so,
in which a value that looks like an exact 5 (e.g., 0.5) is stored with some "decimal dust"
(e.g., looks like 0.5 but is really 0.4999999999999999) because, long story short,
computers are weird.

Would not solve bit flips caused by cosmic rays, though. Which is a very rare, but real thing. Look it up.
}
