% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refuns.R
\name{isid}
\alias{isid}
\title{A Stata-esque function for asserting an id variable is unique.}
\usage{
isid(x)
}
\arguments{
\item{x}{Input the ID to be checked for duplicates. Can be a single variable, or multiple.}
}
\description{
Based on the Stata function of the same name, this function asserts whether rows of a dataframe are unique, and if they are not, it will throw an error and, when sourced, will stop the sourced script code from running any further.This function converts character and factor variables to numeric, or leaves the variable as is if it is already numeric. If the variable is some other type, the function will throw an error and, if the script is being sourced, stop the script from running further.
}
